modify medline writer to use IndexSink, add TupleSourceOperator
