+++ textdb/textdb-dataflow/src/main/java/edu/uci/ics/textdb/dataflow/sink/IndexSink.java
@@ -22,6 +22,7 @@ public class IndexSink extends AbstractSink {
 	public void open() throws Exception {
 		super.open();
 		this.dataWriter.open();
+		this.dataWriter.clearData();
 	}
 
     protected void processOneTuple(ITuple nextTuple) throws StorageException {
