+++ textdb/textdb-perftest/src/main/java/edu/uci/ics/textdb/perftest/medline/MedlineIndexWriter.java
@@ -1,62 +1,86 @@
 package edu.uci.ics.textdb.perftest.medline;
 
+import java.io.File;
 import java.io.FileNotFoundException;
+import java.text.ParseException;
+import java.util.ArrayList;
+import java.util.Scanner;
 
 import org.apache.lucene.analysis.Analyzer;
+import org.json.JSONException;
+import org.json.JSONObject;
 
+import edu.uci.ics.textdb.api.common.Attribute;
+import edu.uci.ics.textdb.api.common.FieldType;
+import edu.uci.ics.textdb.api.common.IField;
 import edu.uci.ics.textdb.api.common.ITuple;
+import edu.uci.ics.textdb.api.common.Schema;
+import edu.uci.ics.textdb.api.dataflow.ISink;
+import edu.uci.ics.textdb.api.dataflow.ISourceOperator;
 import edu.uci.ics.textdb.api.storage.IDataStore;
-import edu.uci.ics.textdb.common.exception.StorageException;
-import edu.uci.ics.textdb.storage.writer.DataWriter;
+import edu.uci.ics.textdb.common.field.DataTuple;
+import edu.uci.ics.textdb.common.utils.Utils;
+import edu.uci.ics.textdb.dataflow.sink.IndexSink;
+import edu.uci.ics.textdb.dataflow.source.FileSourceOperator;
 
 /*
- * This class provides a helper function
- * that can index Medline data.
+ * This class defines Medline data schema.
+ * It also provides functions to read Medline file 
+ * from local machine, parse json format,
+ * and return a generated ITuple one by one.
  * 
  * @author Zuozhi Wang
+ * @author Jinggang Diao
  */
 public class MedlineIndexWriter {
+
+	public static final String PMID = "pmid";
+	public static final String AFFILIATION = "affiliation";
+	public static final String ARTICLE_TITLE = "article_title";
+	public static final String AUTHORS = "authors";
+	public static final String JOURNAL_ISSUE = "journal_issue";
+	public static final String JOURNAL_TITLE = "journal_title";
+	public static final String KEYWORDS = "keywords";
+	public static final String MESH_HEADINGS = "mesh_headings";
+	public static final String ABSTRACT = "abstract";
+	public static final String ZIPF_SCORE = "zipf_score";
+
+	public static final Attribute PMID_ATTR = new Attribute(PMID, FieldType.INTEGER);
+	public static final Attribute AFFILIATION_ATTR = new Attribute(AFFILIATION, FieldType.TEXT);
+	public static final Attribute ARTICLE_TITLE_ATTR = new Attribute(ARTICLE_TITLE, FieldType.TEXT);
+	public static final Attribute AUTHORS_ATTR = new Attribute(AUTHORS, FieldType.TEXT);
+	public static final Attribute JOURNAL_ISSUE_ATTR = new Attribute(JOURNAL_ISSUE, FieldType.STRING);
+	public static final Attribute JOURNAL_TITLE_ATTR = new Attribute(JOURNAL_TITLE, FieldType.TEXT);
+	public static final Attribute KEYWORDS_ATTR = new Attribute(KEYWORDS, FieldType.TEXT);
+	public static final Attribute MESH_HEADINGS_ATTR = new Attribute(MESH_HEADINGS, FieldType.TEXT);
+	public static final Attribute ABSTRACT_ATTR = new Attribute(ABSTRACT, FieldType.TEXT);
+	public static final Attribute ZIPF_SCORE_ATTR = new Attribute(ZIPF_SCORE, FieldType.DOUBLE);
+
+	public static final Attribute[] ATTRIBUTES_MEDLINE = { 
+		PMID_ATTR, AFFILIATION_ATTR, ARTICLE_TITLE_ATTR, AUTHORS_ATTR, JOURNAL_ISSUE_ATTR,
+		JOURNAL_TITLE_ATTR, KEYWORDS_ATTR, MESH_HEADINGS_ATTR, ABSTRACT_ATTR, ZIPF_SCORE_ATTR };
 	
-	/*
-	 * Write medline records from "medlineFilePath"
-	 * to index in "indexPath".
-	 */
-	public static void writeMedlineToIndex(
-			String filePath, IDataStore dataStore, Analyzer luceneAnalyzer) 
-			throws FileNotFoundException, StorageException {
-		writeMedlineToIndex(filePath, dataStore, luceneAnalyzer, Integer.MAX_VALUE);
-	}
+	public static final Schema SCHEMA_MEDLINE = new Schema(ATTRIBUTES_MEDLINE);
 
-	/*
-	 * Write a maximum of "maxDocNumber" records
-	 * from "medlineFilePath"
-	 * to index in "indexPath".
-	 */
-	public static void writeMedlineToIndex(
-			String filePath, IDataStore dataStore, Analyzer luceneAnalyzer, int maxDocNumber) 
-			throws FileNotFoundException, StorageException {
-
-		DataWriter dataWriter = new DataWriter(dataStore, luceneAnalyzer);
-		dataWriter.clearData();
-		dataWriter.open();
-		
-		MedlineReader.open(filePath);
-		
-		int counter = 0;
-		ITuple tuple = null;
-		while ((tuple = MedlineReader.getNextTuple()) != null 
-				&& counter < maxDocNumber) {
-			try {
-				dataWriter.writeTuple(tuple);
-				counter++;
-			} catch (Exception e) {		
-					e.printStackTrace();
-			}
+	
+	private static ITuple recordToTuple(String record) throws JSONException, ParseException {
+		JSONObject json = new JSONObject(record);
+		ArrayList<IField> fieldList = new ArrayList<IField>();
+		for (Attribute attr: ATTRIBUTES_MEDLINE) {
+			fieldList.add(Utils.getField(attr.getFieldType(), json.get(attr.getFieldName()).toString()) );
 		}
-		
-		MedlineReader.close();
-		dataWriter.close();
-
+		IField[] fieldArray = new IField[fieldList.size()];
+		ITuple tuple = new DataTuple(SCHEMA_MEDLINE, fieldList.toArray(fieldArray));
+		return tuple;
 	}
 	
+	
+	public static void writeMedlineToIndex(String filePath, IDataStore dataStore, Analyzer luceneAnalyzer) throws Exception {
+		ISourceOperator fileSourceOperator  = new FileSourceOperator(filePath, (s -> recordToTuple(s)));
+		ISink indexSink = new IndexSink(fileSourceOperator, dataStore.getDataDirectory(), dataStore.getSchema(), luceneAnalyzer);
+		indexSink.open();
+		indexSink.processTuples();
+		indexSink.close();
+	}
+
 }
